backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-groq
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pypdf
python-multipart
python-dotenv
pandas
openpyxl
python-docx
python-docx
pdfplumber
langchain-google-genai
google-generativeai
langchain-openai
a2wsgi
